+++ src/app/core/services/user/user.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from '@angular/core';
+// import { NewUserDTO } from 'src/app/typings';
+// import { HttpClient } from '@angular/common/http';
+@Injectable({
+  providedIn: 'root'
+})
+export class UserService {
+
+  constructor(
+    // private http: HttpClient
+  ) { }
+}