+++ src/app/typings/user/user.interfaces.ts
@@ -1,0 +1,5 @@
+export interface NewUserDTO {
+  email: string;
+  nick: string;
+  password: string;
+}